keep iterator_adaptor self-contained
